secure_file_sender/requirements.txt
prompt_toolkit==3.0.51
pycryptodome==3.22.0
pyOpenSSL==25.0.0
pyOpenSSL==25.0.0
Twisted==24.11.0
